ported over 0.7.0b & c changes, updated version info

+++ core/src/com/shatteredpixel/shatteredpixeldungeon/actors/Char.java
@@ -495,7 +495,7 @@ public abstract class Char extends Actor {
 		}
 	}
 	
-	public int stealth() {
+	public float stealth() {
 		return 0;
 	}
